Renames agent to front, support
